Breakout table column components to their own files
